requirements.txt
flask
flask-cors
grpcio
grpcio-tools
protobuf
pymongo
python-dotenv
requests

flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
transbank-sdk==4.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
Werkzeug==2.3.7 